initial commit of latest intel fork

+++ mlkaps/sample_collection/subprocess_harness.py
@@ -14,7 +14,8 @@ import signal
 import logging
 import pathlib
 from collections import namedtuple
-
+from typing import Dict
+import sys
 
 @dataclass
 class ProcessResult:
@@ -50,9 +51,11 @@ class ProcessCleanupHandler:
         :return: The stdout and stderr of the process
         :rtype: tuple[str, str]
         """
-
-        with suppress(ProcessLookupError):
-            os.killpg(os.getpgid(process.pid), signal.SIGTERM)
+        if os.name == "nt":
+            subprocess.call(['taskkill', '/F', '/T', '/PID',  str(process.pid)])
+        else:                
+            with suppress(ProcessLookupError):
+                os.killpg(os.getpgid(process.pid), signal.SIGTERM)
         return process.communicate()
 
     def run(self, *args, **kwargs) -> ProcessResult:
@@ -61,6 +64,21 @@ class ProcessCleanupHandler:
         :return: The process handle returned by Popen, the stdout and stderr of the process
         :rtype: tuple[subprocess.Popen, str, str]
         """
+
+        # On Windows, .py file are not executable.  
+        # Add "python" as the command. sys.executable is the path to the python interpreter
+
+        if os.name == "nt":
+            assert isinstance(args,tuple)
+            assert len(args) == 1
+            assert isinstance(args[0],list)
+            alist = args[0]
+            wpath = pathlib.WindowsPath(alist[0])
+            if not pathlib.Path(wpath).exists():
+                raise(FileNotFoundError(f"File {wpath} not found"))
+               
+            if wpath.suffix == ".py":
+                args = ([sys.executable] + alist,) 
 
         timed_out = False
         process = None
@@ -99,11 +117,13 @@ class MonoSubprocessHarness:
     def __init__(
         self,
         objectives: list[str],
+        objectives_bounds: Dict[str, float],
         executable_path: pathlib.Path,
         arguments_order: list[str],
         timeout: float | None = None,
     ):
         self.objectives = objectives
+        self.objectives_bounds= objectives_bounds
         self.executable_path = executable_path
         self.timeout = timeout
         self.arguments_order = arguments_order
@@ -144,20 +164,29 @@ class MonoSubprocessHarness:
 
         arguments = [str(sample[k]) for k in self.arguments_order]
         result = self._run_process(arguments)
-
+        # ----- Set to arbitrary value
+        #UB = {obj: bound for obj, bound in zip(self.objectives, self.bounds)}
+        objectives = None
         if result.timed_out:
-            error = f"Process timed out (max {self.timeout} seconds)\n"
+            TO_error = f"Process timed out (max {self.timeout} seconds)\n"
+            #---- save objective to UB, make sense if objective is a time, would require more thinking for true integration.... 
+            error = None
+            objectives = {o: self.objectives_bounds[o] for o in self.objectives} #UB[o]
         elif result.exitcode != 0:
             error = f"Process exited with code {result.exitcode}\n"
+             #---- save objective to UB, make sense if objective is a time, would require more thinking for true integration.... 
+            objectives = {o: self.objectives_bounds[o] for o in self.objectives} #UB[o]
         else:
             objectives, error = self._parse_output(result.stdout)
 
         if error is not None:
-            objectives = {o: float("nan") for o in self.objectives}
+            #---- save objective to UB, make sense if objective is a time, would require more thinking for true integration....
+            objectives = {o: self.objectives_bounds[o] for o in self.objectives}
+                    #float("nan") for o in self.objectives}
             msg = textwrap.indent(f"Arguments: {result.arguments}", "\t| ")
             msg += textwrap.indent(f"\nPID: {result.pid}", "\t| ")
             msg += textwrap.indent(f"\nStdout:\n{result.stdout}", "\t> ")
             error += msg
-
+        
         res_type = namedtuple("SubprocessRunnerOutput", ["data", "error", "timed_out"])
         return res_type(data=objectives, error=error, timed_out=result.timed_out)
